pyproject.toml
[project]
name = "tmdb-service"
version = "0.4.5"
description = "TMDB Service"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.11.18",
    "aiofiles==24.1.0",
    "aiocron==2.1",
    "python-dotenv==1.1.0",
    "psycopg2-binary==2.9.10",
    "sqlalchemy==2.0.41",
]

[tool.ruff]
line-length = 88

[tool.basedpyright]
include = ["tmdb_service"]
typeCheckingMode = "standard"

[tool.setuptools.package-data]
"tmdb_service.tmdb_to_csv" = ["sql/*.sql"]
